+++ src/snap_estimate.js
@@ -95,6 +95,7 @@ export class SnapEstimate {
         this.state_or_territory = inputs.state_or_territory;
         this.monthly_job_income = inputs.monthly_job_income;
         this.monthly_non_job_income = inputs.monthly_non_job_income;
+        this.unemployment_benefits = inputs.unemployment_benefits;
         this.household_size = inputs.household_size;
         this.household_includes_elderly_or_disabled = inputs.household_includes_elderly_or_disabled;
         this.dependent_care_costs = inputs.dependent_care_costs;
@@ -237,6 +238,7 @@ export class SnapEstimate {
             'monthly_non_job_income': this.monthly_non_job_income,
             'court_ordered_child_support_payments': this.court_ordered_child_support_payments,
             'child_support_payments_treatment': this.child_support_payments_treatment,
+            'unemployment_benefits' : this.unemployment_benefits,
         }).calculate();
     }